+++ test/Treasury.t.sol
@@ -2,25 +2,25 @@
 
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/mocks/ERC20Mock.sol";
+import '@openzeppelin/contracts/mocks/ERC20Mock.sol';
 
-import "./mocks/erc/ERC721Mock.sol";
-import "./mocks/erc/ERC1155Mock.sol";
-import "./mocks/PricingExecutor.sol";
+import './mocks/erc/ERC721Mock.sol';
+import './mocks/erc/ERC1155Mock.sol';
+import './mocks/PricingExecutor.sol';
 
-import "../src/contracts/collections/CollectionRegistry.sol";
-import {veFLOOR} from "../src/contracts/tokens/VeFloor.sol";
-import "../src/contracts/tokens/Floor.sol";
-import "../src/contracts/strategies/StrategyRegistry.sol";
-import "../src/contracts/vaults/Vault.sol";
-import "../src/contracts/vaults/VaultFactory.sol";
-import "../src/contracts/RewardsLedger.sol";
-import "../src/contracts/Treasury.sol";
+import '../src/contracts/collections/CollectionRegistry.sol';
+import {veFLOOR} from '../src/contracts/tokens/VeFloor.sol';
+import '../src/contracts/tokens/Floor.sol';
+import '../src/contracts/strategies/StrategyRegistry.sol';
+import '../src/contracts/vaults/Vault.sol';
+import '../src/contracts/vaults/VaultFactory.sol';
+import '../src/contracts/RewardsLedger.sol';
+import '../src/contracts/Treasury.sol';
 
-import "../src/interfaces/vaults/Vault.sol";
-import "../src/interfaces/voting/GaugeWeightVote.sol";
+import '../src/interfaces/vaults/Vault.sol';
+import '../src/interfaces/voting/GaugeWeightVote.sol';
 
-import "./utilities/Environments.sol";
+import './utilities/Environments.sol';
 
 contract TreasuryTest is FloorTest {
     address VAULT_FACTORY = address(10);
@@ -98,7 +98,7 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {FloorMinted}.
      */
-    function test_CanMintFloor(uint256 amount) public {
+    function test_CanMintFloor(uint amount) public {
         vm.assume(amount > 0);
 
         treasury.mint(amount);
@@ -111,7 +111,7 @@ contract TreasuryTest is FloorTest {
      * This should not emit {FloorMinted}.
      */
     function test_CannotMintFloorWithoutPermissions() public {
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(alice);
         treasury.mint(100 ether);
 
@@ -125,7 +125,7 @@ contract TreasuryTest is FloorTest {
      * This should not emit {FloorMinted}.
      */
     function test_CannotMintZeroFloor() public {
-        vm.expectRevert("Cannot mint zero Floor");
+        vm.expectRevert('Cannot mint zero Floor');
         treasury.mint(0);
 
         assertEq(floor.balanceOf(address(treasury)), 0);
@@ -136,7 +136,7 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {Deposit}.
      */
-    function test_CanDepositNativeToken(uint256 amount) public {
+    function test_CanDepositNativeToken(uint amount) public {
         vm.assume(amount <= address(alice).balance);
 
         // Confirm that the {Treasury} starts with 0ETH
@@ -144,7 +144,7 @@ contract TreasuryTest is FloorTest {
 
         // Send the {Treasury} 10ETH from Alice
         vm.prank(alice);
-        (bool sent,) = address(treasury).call{value: amount}("");
+        (bool sent,) = address(treasury).call{value: amount}('');
 
         // Confirm that the ETH was sent successfully
         assertTrue(sent);
@@ -159,7 +159,7 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {DepositERC20}.
      */
-    function test_CanDepositERC20(uint256 mintAmount, uint256 depositAmount) public {
+    function test_CanDepositERC20(uint mintAmount, uint depositAmount) public {
         // The deposit amount must be <= the mint amount
         vm.assume(depositAmount <= mintAmount);
 
@@ -187,7 +187,7 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {DepositERC721}.
      */
-    function test_CanDepositERC721(uint256 tokenId) public {
+    function test_CanDepositERC721(uint tokenId) public {
         // Give Alice an ERC721 to facilitate the test
         erc721.mint(alice, tokenId);
 
@@ -210,12 +210,12 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {DepositERC1155}.
      */
-    function test_CanDepositERC1155(uint256 tokenId, uint256 mintAmount, uint256 depositAmount) public {
+    function test_CanDepositERC1155(uint tokenId, uint mintAmount, uint depositAmount) public {
         // The deposit amount must be <= the mint amount
         vm.assume(depositAmount <= mintAmount);
 
         // Give Alice an ERC1155 to facilitate the test
-        erc1155.mint(alice, tokenId, mintAmount, "");
+        erc1155.mint(alice, tokenId, mintAmount, '');
 
         // Confirm our starting owner of the ERC1155
         assertEq(erc1155.balanceOf(alice, tokenId), mintAmount);
@@ -237,18 +237,18 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {Withdraw}.
      */
-    function test_CanWithdrawNativeToken(uint256 depositAmount, uint256 withdrawAmount) public {
+    function test_CanWithdrawNativeToken(uint depositAmount, uint withdrawAmount) public {
         // The deposit amount must be >= the withdraw amount, but less than Alice's ETH balance
         vm.assume(depositAmount >= withdrawAmount);
         vm.assume(depositAmount <= address(alice).balance);
 
         // Capture Alice's starting ETH balance
-        uint256 aliceStartAmount = address(alice).balance;
-        uint256 bobStartAmount = address(bob).balance;
+        uint aliceStartAmount = address(alice).balance;
+        uint bobStartAmount = address(bob).balance;
 
         // Send the {Treasury} 10ETH from Alice
         vm.prank(alice);
-        (bool sent,) = address(treasury).call{value: depositAmount}("");
+        (bool sent,) = address(treasury).call{value: depositAmount}('');
         assertTrue(sent);
 
         // Confirm that the {Treasury} has received the expected 10ETH
@@ -271,10 +271,10 @@ contract TreasuryTest is FloorTest {
     function test_CannotWithdrawNativeTokenWithoutPermissions() public {
         // Send the {Treasury} 10ETH from Alice
         vm.prank(alice);
-        (bool sent,) = address(treasury).call{value: 10 ether}("");
+        (bool sent,) = address(treasury).call{value: 10 ether}('');
         assertTrue(sent);
 
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(carol);
         treasury.withdraw(carol, 5 ether);
     }
@@ -316,7 +316,7 @@ contract TreasuryTest is FloorTest {
      * This should not emit {WithdrawERC20}.
      */
     function test_CannotWithdrawInvalidERC20() public {
-        vm.expectRevert("ERC20: transfer amount exceeds balance");
+        vm.expectRevert('ERC20: transfer amount exceeds balance');
         vm.prank(bob);
         treasury.withdrawERC20(bob, address(erc20), 3 ether);
     }
@@ -337,7 +337,7 @@ contract TreasuryTest is FloorTest {
         treasury.depositERC20(address(erc20), 4 ether);
         vm.stopPrank();
 
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(carol);
         treasury.withdrawERC20(bob, address(erc20), 3 ether);
     }
@@ -348,7 +348,7 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {WithdrawERC721}.
      */
-    function test_CanWithdrawERC721(uint256 tokenId) public {
+    function test_CanWithdrawERC721(uint tokenId) public {
         // Give Alice an ERC721 to facilitate the test
         erc721.mint(alice, tokenId);
 
@@ -374,7 +374,7 @@ contract TreasuryTest is FloorTest {
      */
     function test_CannotWithdrawInvalidERC721() public {
         // Withdraw the ERC721 to Bob's wallet
-        vm.expectRevert("ERC721: invalid token ID");
+        vm.expectRevert('ERC721: invalid token ID');
         vm.prank(bob);
         treasury.withdrawERC721(bob, address(erc721), 123);
     }
@@ -387,7 +387,7 @@ contract TreasuryTest is FloorTest {
      */
     function test_CannotWithdrawERC721WithoutPermissions() public {
         // Withdraw the ERC721 to Carol's wallet
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(carol);
         treasury.withdrawERC721(bob, address(erc721), 123);
     }
@@ -398,12 +398,12 @@ contract TreasuryTest is FloorTest {
      *
      * This should emit {WithdrawERC721}.
      */
-    function test_CanWithdrawERC1155(uint256 tokenId, uint256 depositAmount, uint256 withdrawAmount) public {
+    function test_CanWithdrawERC1155(uint tokenId, uint depositAmount, uint withdrawAmount) public {
         // The deposit amount must be >= the withdraw amount
         vm.assume(depositAmount >= withdrawAmount);
 
         // Give Alice an ERC1155 to facilitate the test
-        erc1155.mint(alice, tokenId, depositAmount, "");
+        erc1155.mint(alice, tokenId, depositAmount, '');
 
         // Send the {Treasury} tokens from Alice
         vm.startPrank(alice);
@@ -429,7 +429,7 @@ contract TreasuryTest is FloorTest {
      */
     function test_CannotWithdrawInvalidERC1155() public {
         // Withdraw the ERC1155 to Bob's wallet
-        vm.expectRevert("ERC1155: insufficient balance for transfer");
+        vm.expectRevert('ERC1155: insufficient balance for transfer');
         vm.prank(bob);
         treasury.withdrawERC1155(bob, address(erc1155), 1, 1);
     }
@@ -442,7 +442,7 @@ contract TreasuryTest is FloorTest {
      */
     function test_CannotWithdrawInvalidERC1155Amount() public {
         // Give Alice an ERC1155 to facilitate the test
-        erc1155.mint(alice, 1, 3, "");
+        erc1155.mint(alice, 1, 3, '');
 
         // Send the {Treasury} 2 tokens from Alice
         vm.startPrank(alice);
@@ -451,7 +451,7 @@ contract TreasuryTest is FloorTest {
         vm.stopPrank();
 
         // Withdraw the ERC1155 to Bob's wallet
-        vm.expectRevert("ERC1155: insufficient balance for transfer");
+        vm.expectRevert('ERC1155: insufficient balance for transfer');
         vm.prank(bob);
         treasury.withdrawERC1155(bob, address(erc1155), 1, 3);
     }
@@ -464,7 +464,7 @@ contract TreasuryTest is FloorTest {
      */
     function test_CannotWithdrawERC1155WithoutPermissions() public {
         // Give Alice an ERC1155 to facilitate the test
-        erc1155.mint(alice, 1, 3, "");
+        erc1155.mint(alice, 1, 3, '');
 
         // Send the {Treasury} 2 tokens from Alice
         vm.startPrank(alice);
@@ -473,7 +473,7 @@ contract TreasuryTest is FloorTest {
         vm.stopPrank();
 
         // Withdraw the ERC1155 to Bob's wallet
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(carol);
         treasury.withdrawERC1155(bob, address(erc1155), 1, 3);
     }
@@ -494,7 +494,7 @@ contract TreasuryTest is FloorTest {
      * don't pass a `NULL` address value. We expect a revert.
      */
     function test_CannotSetRewardsLedgerContractNullValue() public {
-        vm.expectRevert("Cannot set to null address");
+        vm.expectRevert('Cannot set to null address');
         treasury.setRewardsLedgerContract(address(0));
     }
 
@@ -504,7 +504,7 @@ contract TreasuryTest is FloorTest {
      * be reverted.
      */
     function test_CannotSetRewardsLedgerContractWithoutPermissions() public {
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(alice);
         treasury.setRewardsLedgerContract(address(1));
     }
@@ -520,12 +520,12 @@ contract TreasuryTest is FloorTest {
     }
 
     function test_CannotSetGaugeWeightVoteContractNullValue() public {
-        vm.expectRevert("Cannot set to null address");
+        vm.expectRevert('Cannot set to null address');
         treasury.setGaugeWeightVoteContract(address(0));
     }
 
     function test_CannotSetGaugeWeightVoteContractWithoutPermissions() public {
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(alice);
         treasury.setGaugeWeightVoteContract(address(1));
     }
@@ -533,7 +533,7 @@ contract TreasuryTest is FloorTest {
     /**
      * Retained Treasury Yield Percentage get/set.
      */
-    function test_CanSetRetainedTreasuryYieldPercentage(uint256 percentage) public {
+    function test_CanSetRetainedTreasuryYieldPercentage(uint percentage) public {
         vm.assume(percentage <= 10000);
 
         assertEq(treasury.retainedTreasuryYieldPercentage(), 0);
@@ -542,16 +542,16 @@ contract TreasuryTest is FloorTest {
         assertEq(treasury.retainedTreasuryYieldPercentage(), percentage);
     }
 
-    function test_CannotSetRetainedTreasuryYieldPercentageOverOneHundredPercent(uint256 percentage) public {
+    function test_CannotSetRetainedTreasuryYieldPercentageOverOneHundredPercent(uint percentage) public {
         // Ensure our test amount is over 100%
         vm.assume(percentage > 10000);
 
-        vm.expectRevert("Percentage too high");
+        vm.expectRevert('Percentage too high');
         treasury.setRetainedTreasuryYieldPercentage(percentage);
     }
 
     function test_CannotSetRetainedTreasuryYieldPercentageWithoutPermissions() public {
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(alice);
         treasury.setRetainedTreasuryYieldPercentage(0);
     }
@@ -594,7 +594,7 @@ contract TreasuryTest is FloorTest {
      * anything and we should have our call reverted.
      */
     function test_CannotGetFloorPricesWithoutPricingExecutor() public {
-        vm.expectRevert("No pricing executor set");
+        vm.expectRevert('No pricing executor set');
         treasury.getCollectionFloorPrices();
     }
 
@@ -609,12 +609,12 @@ contract TreasuryTest is FloorTest {
     }
 
     function test_CannotSetPricingExecutorNullValue() public {
-        vm.expectRevert("Cannot set to null address");
+        vm.expectRevert('Cannot set to null address');
         treasury.setPricingExecutor(address(0));
     }
 
     function test_CannotSetPricingExecutorWithoutPermissions() public {
-        vm.expectRevert("Account does not have role");
+        vm.expectRevert('Account does not have role');
         vm.prank(alice);
         treasury.setPricingExecutor(address(1));
     }
@@ -655,11 +655,11 @@ contract TreasuryTest is FloorTest {
 
         // Mock our rewards yield claim amount. For simplicity of future calculations, I've made
         // these the same as the number of users that have (mock) staked against to it
-        vm.mockCall(vaults[0], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint256(8 ether)));
-        vm.mockCall(vaults[1], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint256(1 ether)));
-        vm.mockCall(vaults[2], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint256(5 ether)));
-        vm.mockCall(vaults[3], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint256(6 ether)));
-        vm.mockCall(vaults[4], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint256(4 ether)));
+        vm.mockCall(vaults[0], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint(8 ether)));
+        vm.mockCall(vaults[1], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint(1 ether)));
+        vm.mockCall(vaults[2], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint(5 ether)));
+        vm.mockCall(vaults[3], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint(6 ether)));
+        vm.mockCall(vaults[4], abi.encodeWithSelector(Vault.claimRewards.selector), abi.encode(uint(4 ether)));
 
         // Mock our collection for each vault
         vm.mockCall(vaults[0], abi.encodeWithSelector(IVault.collection.selector), abi.encode(address(1)));
@@ -670,7 +670,7 @@ contract TreasuryTest is FloorTest {
 
         // Mock our vault shares
         address[] memory shareUsers = new address[](7);
-        uint256[] memory userShares = new uint[](7);
+        uint[] memory userShares = new uint[](7);
         (shareUsers[0], shareUsers[1], shareUsers[2], shareUsers[3], shareUsers[4], shareUsers[5], shareUsers[6]) =
             (address(treasury), users[0], users[1], users[2], users[3], users[4], users[5]);
         (userShares[0], userShares[1], userShares[2], userShares[3], userShares[4], userShares[5], userShares[6]) =
@@ -705,11 +705,11 @@ contract TreasuryTest is FloorTest {
         vm.mockCall(vaults[4], abi.encodeWithSelector(Vault.shares.selector), abi.encode(shareUsers, userShares));
 
         // Mock vault share recalculation (ignore)
-        vm.mockCall(vaults[0], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(""));
-        vm.mockCall(vaults[1], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(""));
-        vm.mockCall(vaults[2], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(""));
-        vm.mockCall(vaults[3], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(""));
-        vm.mockCall(vaults[4], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(""));
+        vm.mockCall(vaults[0], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(''));
+        vm.mockCall(vaults[1], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(''));
+        vm.mockCall(vaults[2], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(''));
+        vm.mockCall(vaults[3], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(''));
+        vm.mockCall(vaults[4], abi.encodeWithSelector(Vault.recalculateVaultShare.selector), abi.encode(''));
 
         // Mock our vote distribution
         shareUsers = new address[](5);
@@ -759,7 +759,7 @@ contract TreasuryTest is FloorTest {
 
         // Calling the epoch again should result in a reversion as we have not
         // respected the enforced timelock.
-        vm.expectRevert("Not enough time since last epoch");
+        vm.expectRevert('Not enough time since last epoch');
         treasury.endEpoch();
 
         // After moving forwards 7 days, we can now successfully end another epoch
